shanoir-issue#1466: rename component + fix some merge issues
